+++ bake/azext_bake/_params.py
@@ -40,6 +40,10 @@ def load_arguments(self, _):
         help='Name or ID of a Azure Compute Gallery. You can configure the default using `az configure --defaults bake-gallery=<id>`'
     )
 
+    # repository_path_type = CLIArgumentType(
+    #     options_list=['--repository', '-repo'], type=file_type, validator=repository_path_validator,
+    #     help='Path to the locally cloned repository.')
+
     # yaml_outfile_type validator also validates yaml_outdir_type and yaml_stdout_type
     yaml_outfile_type = CLIArgumentType(options_list=['--outfile'], completer=FilesCompleter(), validator=yaml_out_validator, help='When set, saves the output as the specified file path.')
     yaml_outdir_type = CLIArgumentType(options_list=['--outdir'], completer=DirectoriesCompleter(), help='When set, saves the output at the specified directory.')
@@ -63,9 +67,8 @@ def load_arguments(self, _):
                    'For example if Contoso-Images is provided, the key vault, storage account, and vnet '
                    'will be named Contoso-Images-kv, contosoimagesstorage, and contoso-images-vent respectively.')
         c.argument('principal_id', options_list=['--principal-id', '--principal'],
-                   help='The principal id of a service principal used in the image build pipeline. '
-                   'It will be givin contributor role to sandbox resource group, and the appropriate '
-                   'permissions on the key vault and storage account')
+                   help='The principal id of a service principal used to run az bake from a CI pipeline. '
+                   'It will be givin contributor role to sandbox resource group.')
 
         # c.argument('virtual_network_name', required=False,
         #            options_list=['--vnet-name', '--vnet'],
@@ -119,15 +122,31 @@ def load_arguments(self, _):
         c.ignore('images')
         c.ignore('repo')
 
-    # bake image uses a command level validator, param validators will be ignored
-    with self.argument_context('bake image') as c:
+    with self.argument_context('bake repo setup') as c:
+        c.argument('repository_path', options_list=['--repo-path', '--repo'], type=file_type,
+                   validator=repository_path_validator, help='Path to the locally cloned repository.')
         c.argument('sandbox_resource_group_name', sandbox_resource_group_name_type)
         c.argument('gallery_resource_id', gallery_resource_id_type)
-        c.argument('image_path', options_list=['--image-path', '-i'], type=file_type, help='Path to image to bake.')
-        c.argument('bake_yaml', options_list=['--bake-yaml', '-b'], type=file_type, help='Path to bake.yaml file.')
         c.ignore('sandbox')
         c.ignore('gallery')
-        c.ignore('image')
+
+    # bake image uses a command level validator, param validators will be ignored
+    # with self.argument_context('bake image') as c:
+    #     c.argument('sandbox_resource_group_name', sandbox_resource_group_name_type)
+    #     c.argument('gallery_resource_id', gallery_resource_id_type)
+    #     c.argument('image_path', options_list=['--image-path', '-i'], type=file_type, help='Path to image to bake.')
+    #     c.argument('bake_yaml', options_list=['--bake-yaml', '-b'], type=file_type, help='Path to bake.yaml file.')
+    #     c.ignore('sandbox')
+    #     c.ignore('gallery')
+    #     c.ignore('image')
+
+    with self.argument_context('bake image create') as c:
+        c.argument('image_name', options_list=['--name', '-n'], help='Name of the image to create.')
+        c.argument('repository_path', options_list=['--repo-path', '-r'], type=file_type,
+                   validator=repository_path_validator, help='Path to the locally cloned repository.')
+        # c.argument('outfile', yaml_outfile_type, default='./images/image.yml')
+        # c.argument('outdir', yaml_outdir_type)
+        # c.argument('stdout', yaml_stdout_type)
 
     with self.argument_context('bake yaml export') as c:
         c.argument('sandbox_resource_group_name', sandbox_resource_group_name_type)